Move initialize function for airdrag
