+++ src/disturbances/air_drag.cpp
@@ -7,6 +7,7 @@
 
 #include <cmath>
 #include <environment/global/physical_constants.hpp>
+#include <library/initialize/initialize_file_access.hpp>
 #include <library/math/constants.hpp>
 
 #include "../library/logger/log_utility.hpp"
@@ -90,4 +91,21 @@ std::string AirDrag::GetLogValue() const {
   str_tmp += WriteVector(force_b_N_);
 
   return str_tmp;
+}
+
+AirDrag InitAirDrag(const std::string initialize_file_path, const std::vector<Surface>& surfaces, const Vector<3>& center_of_gravity_b_m) {
+  auto conf = IniAccess(initialize_file_path);
+  const char* section = "AIR_DRAG";
+
+  const double wall_temperature_K = conf.ReadDouble(section, "wall_temperature_degC") + 273.0;
+  const double molecular_temperature_K = conf.ReadDouble(section, "molecular_temperature_degC") + 273.0;
+  const double molecular_weight_g_mol = conf.ReadDouble(section, "molecular_weight_g_mol");
+
+  const bool is_calc_enable = conf.ReadEnable(section, INI_CALC_LABEL);
+  const bool is_log_enable = conf.ReadEnable(section, INI_LOG_LABEL);
+
+  AirDrag air_drag(surfaces, center_of_gravity_b_m, wall_temperature_K, molecular_temperature_K, molecular_weight_g_mol, is_calc_enable);
+  air_drag.is_log_enabled_ = is_log_enable;
+
+  return air_drag;
 }
