+++ src/disturbances/air_drag.hpp
@@ -89,4 +89,13 @@ class AirDrag : public SurfaceForce {
   double CalcFunctionChi(const double s);
 };
 
+/**
+ * @fn InitAirDrag
+ * @brief Initialize AirDrag class
+ * @param [in] initialize_file_path: Initialize file path
+ * @param [in] surfaces: surface information of the spacecraft
+ * @param [in] center_of_gravity_b_m: Center of gravity position vector at body frame [m]
+ */
+AirDrag InitAirDrag(const std::string initialize_file_path, const std::vector<Surface>& surfaces, const Vector<3>& center_of_gravity_b_m);
+
 #endif  // S2E_DISTURBANCES_AIR_DRAG_HPP_
